Add files via upload
separate the templates and routes for admins and users (artist, customers and unregistered users )so that the access of the admin.
I'm not sure whether I need to construct different  model file? Hmmm... it should not be that time and effort consuming as changing the route and templates.
next steps are to deal with the order (transfer order), stocking & delivery system and item system. Notifications are going to be implemented after everything is done? 
hmm... I love the feeling that I have something to do in the winter break...